2007-02-06	Joel Sherrill <joel@OARcorp.com>
	* sp09/screen07.c, sp09/task3.c, sp13/task1.c, sp13/task2.c,
	sp13/task3.c: Address size_t/uint32_t typing issues in message queue
	tests.

+++ testsuites/sptests/sp09/screen07.c
@@ -6,7 +6,7 @@
  *
  *  Output parameters:  NONE
  *
- *  COPYRIGHT (c) 1989-1999.
+ *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
  *  The license and distribution terms for this file may be
@@ -23,7 +23,7 @@
 void Screen7()
 {
   long              buffer[ 4 ];
-  uint32_t    size;
+  size_t      size;
   uint32_t    count;
   rtems_status_code status;
