+++ testsuites/sptests/sp09/task3.c
@@ -9,7 +9,7 @@
  *
  *  Output parameters:  NONE
  *
- *  COPYRIGHT (c) 1989-1999.
+ *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
  *  The license and distribution terms for this file may be
@@ -27,7 +27,7 @@ rtems_task Task_3(
 {
   rtems_status_code status;
   long              buffer[ 4 ];
-  uint32_t    size;
+  size_t            size;
 
   puts( "TA3 - rtems_message_queue_receive - Q 1 - RTEMS_WAIT FOREVER" );
   status = rtems_message_queue_receive(
