+++ testsuites/sptests/sp13/task1.c
@@ -7,7 +7,7 @@
  *
  *  Output parameters:  NONE
  *
- *  COPYRIGHT (c) 1989-1999.
+ *  COPYRIGHT (c) 1989-2007.
  *  On-Line Applications Research Corporation (OAR).
  *
  *  The license and distribution terms for this file may be
@@ -50,8 +50,8 @@ rtems_task Task_1(
   uint32_t          index;
   uint32_t          count;
   rtems_status_code status;
-  uint32_t          size;
-  uint32_t          queue_size;
+  size_t            size;
+  size_t            queue_size;
   unsigned char    *cp;
 
   status = rtems_message_queue_ident(
